pyproject.toml
[project]
name = "qrauthen"
version = "0.1.0"
description = "Implementing Various Features regarding 3.0"
authors = [{ name = "jadaun-abhay", email = "jadaunabhaysingh220@gmail.com" }]
requires-python = ">=3.12,<4.0"
readme = "README.md"
dependencies = [
    "django (==5.1)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "mysqlclient (==2.2.7)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-sslserver-v2 (>=1.0,<2.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "uuid6 (>=2025.0.1,<2026.0.0)",
    "qrcode (>=8.2,<9.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-eventstream (>=5.3.2,<6.0.0)",
    "channels (>=4.3.1,<5.0.0)",
    "pyopenssl (>=25.1.0,<26.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "celery-valkey-backend (>=0.1.0,<0.2.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "svgwrite (>=1.4.3,<2.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "redis (>=6.4.0,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.uv]
package = false
